--- a/app/ante/authz.go
+++ b/app/ante/authz.go
@@ -53,7 +53,7 @@ func (ald AuthzLimiterDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate
 	}
 
 	if err := ald.checkDisabledMsgs(tx.GetMsgs(), false, 0); err != nil {
-		return ctx, errorsmod.Wrapf(errortypes.ErrUnauthorized, "unauthorized error: %s", err.Error())
+		return ctx, errorsmod.Wrapf(errortypes.ErrUnauthorized, "%s", err.Error())
 	}
 	return next(ctx, tx, simulate)
 }