--- a/app/ante/eth.go
+++ b/app/ante/eth.go
@@ -58,7 +58,7 @@ func NewCachedAccountGetter(ctx sdk.Context, ak evmtypes.AccountKeeper) AccountG
 }
 
 // VerifyEthAccount validates checks that the sender balance is greater than the total transaction cost.
-// The account will be created in memory if it doesn't exis, i.e cannot be found on store, which will eventually set to
+// The account will be created in memory if it doesn't exist, i.e cannot be found on store, which will eventually set to
 // store when increasing nonce.
 // This AnteHandler decorator will fail if:
 // - any of the msgs is not a MsgEthereumTx