--- a/x/evm/statedb/state_object.go
+++ b/x/evm/statedb/state_object.go
@@ -23,7 +23,6 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
 	ethermint "github.com/evmos/ethermint/types"
-	"github.com/evmos/ethermint/x/evm/types"
 )
 
 var emptyCodeHash = crypto.Keccak256(nil)
@@ -44,16 +43,14 @@ func NewEmptyAccount() *Account {
 
 // NewAccountFromSdkAccount extracts the nonce and code hash from the provided SDK account.
 func NewAccountFromSdkAccount(acct sdk.AccountI) *Account {
-	codeHash := types.EmptyCodeHash
-	ethAcct, ok := acct.(ethermint.EthAccountI)
-	if ok {
-		codeHash = ethAcct.GetCodeHash().Bytes()
+	acc := NewEmptyAccount()
+	acc.Nonce = acct.GetSequence()
+
+	if ethAcct, ok := acct.(ethermint.EthAccountI); ok {
+		acc.CodeHash = ethAcct.GetCodeHash().Bytes()
 	}
 
-	return &Account{
-		Nonce:    acct.GetSequence(),
-		CodeHash: codeHash,
-	}
+	return acc
 }
 
 // IsContract returns if the account contains contract code.