reuse account after verify account balance to avoid get again when check sender sequence

benchmark need remove the skip [part](https://github.com/crypto-org-chain/ethermint/blob/7ef8af1388ac62f925275cd9511528281e53be73/app/ante/eth.go#L47-L49)
```
                                │    sec/op    │   sec/op     vs base              │
ERC20Transfer/memiavl-stm-1-12    898.8m ±  8%   847.5m ± 1%  -5.71% (p=0.002 n=6)
ERC20Transfer/memiavl-stm-8-12    185.4m ± 14%   172.4m ± 2%  -7.03% (p=0.004 n=6)
ERC20Transfer/memiavl-stm-16-12   199.1m ±  7%   192.8m ± 3%  -3.17% (p=0.026 n=6)
ERC20Transfer/memiavl-stm-32-12   244.6m ±  7%   246.3m ± 9%       ~ (p=0.818 n=6)
geomean                           300.1m         288.6m       -3.84%
```

## Description

<!-- Add a description of the changes that this PR introduces and the files that
are the most critical to review.
-->

______

For contributor use:

- [ ] Targeted PR against correct branch (see [CONTRIBUTING.md](https://github.com/cosmos/cosmos-sdk/blob/master/CONTRIBUTING.md#pr-targeting))
- [ ] Linked to Github issue with discussion and accepted design OR link to spec that describes this work.
- [ ] Code follows the [module structure standards](https://github.com/cosmos/cosmos-sdk/blob/master/docs/building-modules/structure.md).
- [ ] Wrote unit and integration [tests](https://github.com/cosmos/cosmos-sdk/blob/master/CONTRIBUTING.md#testing)
- [ ] Updated relevant documentation (`docs/`) or specification (`x/<module>/spec/`)
- [ ] Added  relevant `godoc` [comments](https://blog.golang.org/godoc-documenting-go-code).
- [ ] Added a relevant changelog entry to the `Unreleased` section in `CHANGELOG.md`
- [ ] Re-reviewed `Files changed` in the Github PR explorer

______

For admin use:

- [ ] Added appropriate labels to PR (ex. `WIP`, `R4R`, `docs`, etc)
- [ ] Reviewers assigned
- [ ] Squashed all commits, uses message "Merge pull request #XYZ: [title]" ([coding standards](https://github.com/tendermint/coding/blob/master/README.md#merging-a-pr))