@yihuang 2024-08-13
did you do benchmarks for this, for example the `BenchmarkERC20Transfer` in cronos repo.
This change mainly to save the overhead of runtime type casting in `msg.(*MsgEthereumTx)`, right? I wonder how much performance gain there are.