@mmsqe 2024-08-13
> did you do benchmarks for this, for example the `BenchmarkERC20Transfer` in cronos repo. This change mainly to save the overhead of runtime type casting in `msg.(*MsgEthereumTx)`, right? I wonder how much performance gain there are.

yea, seems only small gain
```
ERC20Transfer/memiavl-stm-1-12     1.610 ± 2%    1.601 ± 3%       ~ (p=0.083 n=8)
ERC20Transfer/memiavl-stm-8-12    266.5m ± 6%   264.0m ± 3%       ~ (p=0.328 n=8)
ERC20Transfer/memiavl-stm-16-12   266.4m ± 6%   253.3m ± 0%  -4.94% (p=0.010 n=8)
ERC20Transfer/memiavl-stm-32-12   290.9m ± 4%   289.8m ± 4%       ~ (p=1.000 n=8)
geomean                           427.0m        419.7m       -1.70%
```