@yihuang 2024-08-15
```
goos: darwin
goarch: arm64
pkg: github.com/crypto-org-chain/cronos/v2/app
                                │ /tmp/before.txt │        /tmp/after.txt        │
                                │     sec/op      │   sec/op     vs base         │
ERC20Transfer/memiavl-stm-16-16       123.7m ± 2%   124.1m ± 2%  ~ (p=0.589 n=6)

                                │ /tmp/before.txt │           /tmp/after.txt           │
                                │      B/op       │     B/op      vs base              │
ERC20Transfer/memiavl-stm-16-16      486.5Mi ± 0%   494.8Mi ± 0%  +1.69% (p=0.002 n=6)

                                │ /tmp/before.txt │          /tmp/after.txt           │
                                │    allocs/op    │  allocs/op   vs base              │
ERC20Transfer/memiavl-stm-16-16       6.232M ± 0%   6.267M ± 0%  +0.56% (p=0.004 n=6)
```